requirements.txt
certifi>=2024.0.0
psycopg2-binary==2.9.9
pymongo==4.15.5
python-dotenv==1.0.0
